Move framework tests to respective folders
